% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multinomial_n.R
\name{calculate_all_combinations_internal}
\alias{calculate_all_combinations_internal}
\title{Internal function to calculate all combinations}
\usage{
calculate_all_combinations_internal(
  x,
  sex_categories,
  include_pooled,
  remove_outliers,
  min_proportion,
  max_cv,
  quiet
)
}
\description{
Internal function to calculate all combinations
}
\keyword{internal}
