% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_length_compositions.R
\name{calculate_length_compositions_core}
\alias{calculate_length_compositions_core}
\title{Core calculation function for length compositions}
\usage{
calculate_length_compositions_core(
  fish_data,
  strata_data,
  length_range,
  plus_group,
  minus_group,
  scaling_type,
  lw_params_male,
  lw_params_female,
  lw_params_unsexed,
  lengths,
  strata_names
)
}
\description{
Core calculation function for length compositions
}
\keyword{internal}
