% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multinomial_n.R
\name{calculate_single_combination_internal}
\alias{calculate_single_combination_internal}
\title{Internal function to calculate single combination}
\usage{
calculate_single_combination_internal(
  x,
  stratum,
  sex,
  remove_outliers,
  min_proportion,
  min_cv,
  max_cv,
  trace
)
}
\description{
Internal function to calculate single combination
}
\keyword{internal}
