% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multinomial_n.R
\name{fit_multinomial_model_internal}
\alias{fit_multinomial_model_internal}
\title{Internal function to fit the multinomial model from proportions and CVs}
\usage{
fit_multinomial_model_internal(
  proportions,
  cvs,
  bins,
  remove_outliers,
  min_proportion,
  min_cv,
  max_cv,
  trace,
  analysis_type = "raw_vectors",
  sex = NULL,
  stratum = NULL
)
}
\description{
Internal function to fit the multinomial model from proportions and CVs
}
\keyword{internal}
