% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_bias_correction.R
\name{length_bias_correction}
\alias{length_bias_correction}
\title{Length Measurement Bias Correction Functions}
\description{
Functions to detect and correct length measurement bias (rounding/heaping)
in fisheries data, particularly preference for lengths divisible by 5.
}
\author{
Scala Package Team
}
