% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multinomial_n.R
\name{plot.multinomial_n}
\alias{plot.multinomial_n}
\title{Plot method for multinomial_n objects}
\usage{
\method{plot}{multinomial_n}(x, ...)
}
\arguments{
\item{x}{A multinomial_n object from calculate_multinomial_n()}

\item{...}{Additional arguments (not used)}
}
\value{
A ggplot object
}
\description{
Creates a ggplot showing observed CVs versus proportions with the fitted
multinomial curve overlaid. Points are the observed data used in the fit
and the curve shows the theoretical relationship CV = sqrt((1-P) / (n*P))
for the estimated effective sample size n.
}
