% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.age_composition.R
\name{print.age_composition}
\alias{print.age_composition}
\title{Print Method for Age Composition Objects}
\usage{
\method{print}{age_composition}(x, ...)
}
\arguments{
\item{x}{An age_composition object from calculate_age_compositions()}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Age Composition Objects
}
