% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.age_length_key.R
\name{print.age_length_key}
\alias{print.age_length_key}
\title{Print Method for Age-Length Key}
\usage{
\method{print}{age_length_key}(x, ...)
}
\arguments{
\item{x}{An age-length key data frame or list of sex-specific age-length keys}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Age-Length Key
}
