% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cohort_alk.R
\name{print.cohort_alk}
\alias{print.cohort_alk}
\title{Print method for cohort_alk objects}
\usage{
\method{print}{cohort_alk}(x, ...)
}
\arguments{
\item{x}{A cohort_alk object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cohort_alk objects
}
