% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multinomial_alk.R
\name{print.multinomial_alk}
\alias{print.multinomial_alk}
\title{Print a multinomial age-at-length model}
\usage{
\method{print}{multinomial_alk}(x, ...)
}
\arguments{
\item{x}{A multinomial_alk object}

\item{...}{Additional arguments}
}
\description{
Print a multinomial age-at-length model
}
