% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multinomial_n.R
\name{print.multinomial_n}
\alias{print.multinomial_n}
\title{Print method for multinomial_n objects}
\usage{
\method{print}{multinomial_n}(x, ...)
}
\arguments{
\item{x}{A multinomial_n object from calculate_multinomial_n()}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for multinomial_n objects
}
