% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multinomial_n.R
\name{print.multinomial_n_summary}
\alias{print.multinomial_n_summary}
\title{Print method for multinomial_n_summary objects}
\usage{
\method{print}{multinomial_n_summary}(x, ...)
}
\arguments{
\item{x}{A multinomial_n_summary object from calculate_all_multinomial_n()}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for multinomial_n_summary objects
}
