% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ordinal_alk.R
\name{print.ordinal_alk}
\alias{print.ordinal_alk}
\title{Print an ordinal age-at-length model}
\usage{
\method{print}{ordinal_alk}(x, ...)
}
\arguments{
\item{x}{An ordinal_alk object}

\item{...}{Additional arguments}
}
\description{
Print an ordinal age-at-length model
}
