% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weight_age.R
\name{print.weight_age_model}
\alias{print.weight_age_model}
\title{Print method for weight_age_model objects}
\usage{
\method{print}{weight_age_model}(x, ...)
}
\arguments{
\item{x}{A weight_age_model object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for weight_age_model objects
}
