% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age_compositions.R
\name{apply_age_length_key}
\alias{apply_age_length_key}
\title{Apply Age-Length Key to Length Compositions}
\usage{
apply_age_length_key(
  length_comp,
  age_length_key,
  ages,
  lengths,
  plus_group_age = FALSE,
  minus_group_age = FALSE
)
}
\arguments{
\item{length_comp}{3D array of length compositions (length x sex x stratum)}

\item{age_length_key}{Data frame with length, age, proportion columns}

\item{ages}{Vector of age bins}

\item{lengths}{Vector of length bins}

\item{plus_group_age}{Logical, whether to apply plus group for ages}

\item{minus_group_age}{Logical, whether to apply minus group for ages}
}
\value{
3D array of age compositions (age x sex x stratum)
}
\description{
Internal function to convert length compositions to age compositions using age-length key.
}
\keyword{internal}
