% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_commercial_test_data}
\alias{generate_commercial_test_data}
\title{Generate commercial fisheries test data}
\usage{
generate_commercial_test_data()
}
\description{
Generate commercial fisheries test data
}
