% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.r
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate test data for CAA package examples}
\usage{
generate_test_data(data_type = "commercial")
}
\arguments{
\item{data_type}{Character, either "commercial" for weight-based data or "survey" for density-based data}
}
\value{
List containing fish_data and strata_data for testing
}
\description{
Generate test data for CAA package examples
}
\keyword{internal}
