% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ordinal_alk.R
\name{print.ordinal_alk}
\alias{print.ordinal_alk}
\title{Print method for ordinal_alk objects}
\usage{
\method{print}{ordinal_alk}(x, ...)
}
\arguments{
\item{x}{An ordinal_alk object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ordinal_alk objects
}
